+++ src/colonyscanalyser/main.py
@@ -1,19 +1,25 @@
 ﻿# System modules
 import sys
 import argparse
-from typing import Union, Dict, List, Tuple
+from typing import Optional, Union, Dict, List, Tuple
 from pathlib import Path
 from datetime import datetime
 from distutils.util import strtobool
 from collections import defaultdict
 from multiprocessing import Pool, cpu_count
 from functools import partial
+try:
+    from importlib import metadata
+except ImportError:
+    # Running on pre-3.8 Python, use importlib-metadata package
+    import importlib_metadata as metadata
 
 # Third party modules
 from numpy import ndarray, diff
 
 # Local modules
 from colonyscanalyser import (
+    config,
     utilities,
     file_access,
     imaging,
@@ -22,6 +28,46 @@ from colonyscanalyser import (
 from .image_file import ImageFile, ImageFileCollection
 from .plate import Plate, PlateCollection
 from .colony import Colony, timepoints_from_image, colonies_from_timepoints, timepoints_from_image
+
+
+def argparse_init(*args, **kwargs) -> argparse.ArgumentParser:
+    """
+    Initialise an ArgumentParser instance with the standard package arguments
+
+    :param args: positional arguments to pass to the ArgumentParser initialiser
+    :param kwargs: keyword arguments to pass to the ArgumentParser initialiser
+    :returns: an ArgumentParser instance with the standard package arguments
+    """
+    parser = argparse.ArgumentParser(*args, **kwargs)
+    
+    parser.add_argument("path", type = str,
+                        help = "Image files location", default = None)
+    parser.add_argument("-dpi", "--dots_per_inch", type = int, default = config.DOTS_PER_INCH,
+                        help = "The image DPI (dots per inch) setting", metavar = "N")
+    parser.add_argument("--image_formats", default = config.SUPPORTED_FORMATS, action = "version", version = str(config.SUPPORTED_FORMATS),
+                        help = "The supported image formats")
+    parser.add_argument("-mp", "--multiprocessing", type = strtobool, default = config.MULTIPROCESSING,
+                        help = "Enables use of more CPU cores, faster but more resource intensive",  metavar = "BOOLEAN")
+    parser.add_argument("-p", "--plots", type = int, default = config.OUTPUT_PLOTS,
+                        help = "The detail level of plot images to store on disk", metavar = "N")
+    parser.add_argument("--plate_edge_cut", type = int, default = config.PLATE_EDGE_CUT,
+                        help = "The exclusion area from the plate edge, as a percentage of the plate diameter", metavar = "N")
+    parser.add_argument("--plate_labels", type = str, nargs = "*", default = list(),
+                        help = "A list of labels to identify each plate. Plates are ordered from top left, in rows. Example usage: --plate_labels plate1 plate2",
+                        metavar = "LABEL")
+    parser.add_argument("--plate_lattice", type = int, nargs = 2, default = config.PLATE_LATTICE,
+                        help = "The row and column co-ordinate layout of plates. Example usage: --plate_lattice 3 3",
+                        metavar = ("ROW", "COL"))
+    parser.add_argument("--plate_size", type = int, default = config.PLATE_SIZE,
+                        help = "The plate diameter, in millimetres", metavar = "N")
+    parser.add_argument("--use_cached_data", type = strtobool, default = config.USE_CACHED_DATA,
+                        help = "Allow use of previously calculated data", metavar = "BOOLEAN")
+    parser.add_argument("-v", "--verbose", type = int, default = config.OUTPUT_VERBOSE,
+                        help = "Information output level", metavar = "N")
+    parser.add_argument("--version", action = "version", version = f"ColonyScanlayser {metadata.version('colonyscanalyser')}",
+                        help = "The package version number")
+
+    return parser
 
 
 def segment_image(
@@ -115,34 +161,16 @@ def image_file_to_timepoints(
 
 # flake8: noqa: C901
 def main():
-    parser = argparse.ArgumentParser(
+    parser = argparse_init(
         description = "An image analysis tool for measuring microorganism colony growth",
-        formatter_class = argparse.ArgumentDefaultsHelpFormatter
-    )
-    parser.add_argument("path", type = str,
-                        help = "Image files location", default = None)
-    parser.add_argument("-dpi", "--dots_per_inch", type = int, default = 300,
-                        help = "The image DPI (dots per inch) setting")
-    parser.add_argument("-mp", "--multiprocessing", type = strtobool, default = True,
-                        help = "Enables use of more CPU cores, faster but more resource intensive")
-    parser.add_argument("-p", "--plots", type = int, default = 1,
-                        help = "The detail level of plot images to store on disk")
-    parser.add_argument("--plate_edge_cut", type = int, default = 5,
-                        help = "The exclusion area from the plate edge, as a percentage of the plate diameter")
-    parser.add_argument("--plate_labels", type = str, nargs = "*", default = list(),
-                        help = "A list of labels to identify each plate. Plates are ordered from top left, in rows. Example usage: --plate_labels plate1 plate2")
-    parser.add_argument("--plate_lattice", type = int, nargs = 2, default = (3, 2),
-                        metavar = ("ROW", "COL"),
-                        help = "The row and column co-ordinate layout of plates. Example usage: --plate_lattice 3 3")
-    parser.add_argument("--plate_size", type = int, default = 90,
-                        help = "The plate diameter, in millimetres")
-    parser.add_argument("--use_cached_data", type = strtobool, default = False,
-                        help = "Allow use of previously calculated data")
-    parser.add_argument("-v", "--verbose", type = int, default = 1,
-                        help = "Information output level")
+        formatter_class = argparse.ArgumentDefaultsHelpFormatter,
+        usage = "%(prog)s '/image/file/path/' [OPTIONS]"
+        )
 
+    # Retrieve and parse arguments
     args = parser.parse_args()
     BASE_PATH = args.path
+    IMAGE_FORMATS = args.image_formats
     PLOTS = args.plots
     PLATE_LABELS = {plate_id: label for plate_id, label in enumerate(args.plate_labels, start = 1)}
     PLATE_LATTICE = tuple(args.plate_lattice)
@@ -170,13 +198,12 @@ def main():
         print(f"Working directory: {BASE_PATH}")
 
     # Check if processed image data is already stored and can be loaded
-    segmented_image_data_filename = "cached_data"
     plates = None
     if USE_CACHED:
         if VERBOSE >= 1:
             print("Attempting to load cached data")
         plates = file_access.load_file(
-            BASE_PATH.joinpath("data", segmented_image_data_filename),
+            BASE_PATH.joinpath(config.DATA_DIR, config.CACHED_DATA_FILE_NAME),
             file_access.CompressionMethod.LZMA,
             pickle = True
         )
@@ -195,8 +222,7 @@ def main():
 
     if not USE_CACHED or plates is None:
         # Find images in working directory
-        image_formats = ["tif", "tiff", "png", "bmp"]
-        image_paths = file_access.get_files_by_type(BASE_PATH, image_formats)
+        image_paths = file_access.get_files_by_type(BASE_PATH, IMAGE_FORMATS)
 
         # Store images as ImageFile objects
         # Timestamps are automatically read from filenames
@@ -302,24 +328,24 @@ def main():
                 break
 
             plate = plates.get_item(plate_id)
-            plate.items = colonies_from_timepoints(plate_timepoints, distance_tolerance = 2)
+            plate.items = colonies_from_timepoints(plate_timepoints, distance_tolerance = config.COLONY_DISTANCE_MAX)
             if VERBOSE >= 3:
                 print(f"{plate.count} objects located on plate {plate.id}, before filtering")
 
             # Filter colonies to remove noise, background objects and merged colonies
             timestamp_diff_std = diff(image_files.timestamps_elapsed_seconds[1:]).std()
-            timestamp_diff_std += 20
+            timestamp_diff_std += config.COLONY_TIMESTAMP_DIFF_MAX
             plate.items = list(filter(lambda colony:
                 # Remove objects that do not have sufficient data points
-                len(colony.timepoints) > 5 and
+                len(colony.timepoints) > config.COLONY_TIMEPOINTS_MIN and
                 # No colonies should be visible at the start of the experiment
                 colony.time_of_appearance.total_seconds() > 0 and
                 # Remove objects with large gaps in the data
                 diff([t.timestamp.total_seconds() for t in colony.timepoints[1:]]).std() < timestamp_diff_std and
                 # Remove object that do not show growth, these are not colonies
-                colony.timepoint_last.area > 4 * colony.timepoint_first.area and
+                colony.timepoint_last.area > config.COLONY_GROWTH_FACTOR_MIN * colony.timepoint_first.area and
                 # Objects that appear with a large initial area are either merged colonies or noise
-                colony.timepoint_first.area < 10,
+                colony.timepoint_first.area < config.COLONY_FIRST_AREA_MAX,
                 plate.items
             ))
 
@@ -333,8 +359,8 @@ def main():
             sys.exit()
 
     # Store pickled data to allow quick re-use
-    save_path = file_access.create_subdirectory(BASE_PATH, "data")
-    save_path = save_path.joinpath(segmented_image_data_filename)
+    save_path = file_access.create_subdirectory(BASE_PATH, config.DATA_DIR)
+    save_path = save_path.joinpath(config.CACHED_DATA_FILE_NAME)
     save_status = file_access.save_file(save_path, plates, file_access.CompressionMethod.LZMA)
     if VERBOSE >= 1:
         if save_status:
@@ -346,7 +372,7 @@ def main():
     if VERBOSE >= 1:
         print("Saving data to CSV")
         
-    save_path = BASE_PATH.joinpath("data")
+    save_path = BASE_PATH.joinpath(config.DATA_DIR)
     for plate in plates.items:
         for colony in plate.items:
             test = colony.__iter__()
@@ -363,7 +389,7 @@ def main():
     # Can't guarantee that the original images and full list of time points
     # will be available when using cached data
     if image_files is not None:
-        save_path = file_access.create_subdirectory(BASE_PATH, "plots")
+        save_path = file_access.create_subdirectory(BASE_PATH, config.PLOTS_DIR)
         if PLOTS >= 1:
             if VERBOSE >= 1:
                 print("Saving plots")