+++ docs/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.3] - 2020-06-02
+### Added
+- `--version` command line argument to display installed package version number
+- `--image_formats` command line argument to display supported image formats
+- Package configuration and now consolidated and loaded from a settings module
+### Changed
+- Improved the output from the `--help` command line argument to make it a little clearer
+
 ## [0.4.2] - 2020-05-18
 ### Added
 - Support for `bmp` image files